Split primitives l1 module
